修改386编译不过去bug 将int类型修改为int64位

+++ parser.go
@@ -10068,7 +10068,7 @@ yynewstate:
 		{
 			x := types.NewFieldType(mysql.TypeDuration)
 			x.Flen, _ = mysql.GetDefaultFieldLengthAndDecimalForCast(mysql.TypeDuration)
-			x.Decimal = yyS[yypt-0].item.(int)
+			x.Decimal = yyS[yypt-0].item.(int64)
 			if x.Decimal > 0 {
 				x.Flen = x.Flen + 1 + x.Decimal
 			}
